requirements.txt
feedparser==6.0.11
python_dateutil==2.9.0.post0
Requests==2.32.3
